+++ src/main/java/org/fit/cssbox/layout/TableColumn.java
@@ -260,7 +260,7 @@ public class TableColumn extends BlockBox
 	            if (width.isPercentage())
 	            {
 	            	wrelative = true;
-                    percent = (int) Math.round(width.getValue());
+                    percent = Math.round(width.getValue());
                     if (percent == 0)
                         wrelative = false; //consider 0% as absolute 0
 	            }
